modularized the project by rearranging the files
