+++ FFT/CooleyTukeyFFT/readme
@@ -1,0 +1,18 @@
+## Cooley-Tukey FFT methods
+
+*Recommended format*
+
+`void FFT_method(Complex *vec, (unsigned) int N)`
+
+- `Complex`(dtype) $\equiv$ `complex<double>`
+- `N`: length of the array
+
+the FFT result will be stored in the original `vec`
+
+
+## List of Methods Available
+
+#### 1-D FFT
+- the Recursive method
+- the Iterative method
+- the Wrapper function for FFT method from FFTW library
