+++ FFT/data_loader/readme
@@ -1,0 +1,3 @@
+## Data Loader 
+
+Reading data from a specified path (root path at ./Data folder) as an Dataset class object
